Fix font rendering.
OpenJDK fonts are rendered based on integer pixel coordinates. Somewhere
during the calculations, it rounds up when it should round down. This
causes some letters to be one pixel too far, and it makes the
Font.stringWidth() method return the wrong values.

To fix that, we skip the pixel-resolution step. Using GlyphVector, the
text is rendered in floating point, and then that is converted to a
Shape which is then drawn.